v1.3.3 - Create ui_group config nodes. Groups can now be assigned from dropdown bars as with tabs and edited from the config options in Node-RED
NOTE: work still to be done here to allow groups to have custom column widths and to potentially restrict groups to be within a single tab only. Currently you can define two nodes to be in the same group, but if you define a different tab for the nodes, that creates a new instance of the same group on a different tab.